// apps/gateway/src/friend/models/friend-count-aggregate.input.ts
import { Field, InputType } from '@nestjs/graphql';

@InputType()
export class FriendCountAggregateInput {
  @Field(_type => Boolean, { nullable: true })
  id?: true;

  @Field(_type => Boolean, { nullable: true })
  createdAt?: true;

  @Field(_type => Boolean, { nullable: true })
  updatedAt?: true;

  @Field(_type => Boolean, { nullable: true })
  deleted?: true;

  @Field(_type => Boolean, { nullable: true })
  name?: true;

  @Field(_type => Boolean, { nullable: true })
  birthday?: true;

  @Field(_type => Boolean, { nullable: true })
  description?: true;

  @Field(_type => Boolean, { nullable: true })
  _all?: true;
}